avm2: Move method naming to Executable
